backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
python-decouple==3.8
alembic==1.12.1
email-validator==2.1.0
